backend/requirements.txt
chainlit
llama-index
python-dotenv
